Planning: added test for trajectory cost. Seperated comparable cost from trajectory cost.
