+++ modules/planning/tasks/dp_poly_path/BUILD
@@ -9,6 +9,7 @@ cc_library(
         "trajectory_cost.cc",
     ],
     hdrs = [
+        "comparable_cost.h",
         "dp_road_graph.h",
         "trajectory_cost.h",
     ],
@@ -28,7 +29,7 @@ cc_library(
         "//modules/planning/math/curve1d:quintic_polynomial_curve1d",
         "//modules/planning/proto:dp_poly_path_config_proto",
         "//modules/planning/reference_line",
-        "@eigen//:eigen",
+        "@eigen",
     ],
 )
 
@@ -43,6 +44,18 @@ cc_library(
     deps = [
         ":dp_poly_path",
         "//modules/planning/tasks:path_optimizer",
+    ],
+)
+
+cc_test(
+    name = "comparable_cost_test",
+    size = "small",
+    srcs = [
+        "comparable_cost_test.cc",
+    ],
+    deps = [
+        ":dp_poly_path",
+        "@gtest//:main",
     ],
 )
